fix routing, add admin page to routes, add logo
