+++ frontend/src/index.tsx
@@ -2,21 +2,30 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import App from './App';
+import Chatbox from './components/Chatbox';
+import { AdminPage } from './pages/AdminPage';
 
 const domNode = document.getElementById('root')!;
 const root = createRoot(domNode);
 
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call
 const router = createBrowserRouter([
   {
     Component: App,
     children: [
       {
         path: '/',
-        Component: App,
+        Component: Chatbox,
       },
-    ],
-  },
+      {
+        path: 'chat/:id',
+        Component: Chatbox,
+      },
+      {
+        path: 'admin',
+        Component: AdminPage
+      }
+    ]
+  }
 ]);
 
 root.render(
