+++ frontend/types.d.ts
@@ -1,0 +1,4 @@
+declare module '*.png' {
+  const value: string;
+  export default value;
+}
